requirements.txt
aiofiles==24.1.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.3.0
beanie==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
lazy-model==0.3.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rich==14.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
